requirements.txt
algoliasearch-django>=3.0,<4.0
django>=4.0.0,<4.1.0
djangorestframework
pyyaml
requests
django-cors-headers

